Teamviewer add

+++ data/post-install.sh
@@ -1,11 +1,16 @@
 #!/bin/bash
 
 userdel -r drk
-rm /etc/NetworkManager/system-connections/DRK_SAK_Mobile.nmconnection
+
+if test -f /etc/NetworkManager/system-connections/DRK_SAK_Mobile.nmconnection; then
+    rm /etc/NetworkManager/system-connections/DRK_SAK_Mobile.nmconnection
+fi
+
+teamviewer --daemon start &
+
+teamviewer assignment --id
 
 rm "$0"
 
 # systemctl enable displaylink
 # systemctl start displaylink
-# systemctl enable teamviewer
-# teamviewer --daemon start
